Use the conf.{static,shared} instead of passing in the builder.
I ran into one logical problem with passing around abstract builders.
In order to test dynamic loading of code, we need to generate a shared
libary. Unfortunately, the way things are created, we can't test a
loading with a static executable since the test can't actually create
a test shared libary. So, we'll just make it explicit which builder
we're using so we can use either one on demand.