Add a simple package system and another example.
One minor change is that we now only compile one file at a time. This
allows us to run builds on multiple threads at a time, whereas before it
was forced to be linearized. I also added a mechanism to force evaluation
of a future in all the builders. It's a bit of a hack though.