Implement an environment object that will cache all the fbuild state.
Previously, we used a standard dictionary to store all the state of fbuild.
Unfortunately, this meant that we were also hardcoding the paths to all
objects, so it would be impossible to configure two different compilers,
unless they were separated by passing sub-dictionaries around. This works,
but then you don't have the advantage of reusing computations.

This class changes it. It makes the assumption that configuring a builder
is purely dependent on it's imputs. This lets us cache the data, and return
builders if they were configured with the same inputs. This will allow us
to remove the configure phase, and make it implicit in the build.

In order to support multiple calls to a function and caching the different
results, we can't use a dictionary since kwargs are not hashable. So, we
have to store the arguments in a list, and then manually step over each
one to test if the args and kwargs are equal. We're assuming that multiple
builders will be rare, so this shouldn't be expensive.