Cache the results of the tests in the env, and return the value directly.
Rather than looking up tests in a dictionary, it makes more sense to call
the test function, and cache it in the background. This lets us return
the same value if it happens to be called with the same arguments. For
instance, a couple builders need to know what the platform is, so if they
call it with the same arguments, they'll get the same platform without
having to recompute it.