pickle state to a string, then write it out to prevent corrupted state files.
It's possible for pickling to raise an exception, which would result in a
partially written out state file. To prevent this, we'll pickle to a string
first, then write it out. Technically there still could be some corruption,
since we could run into some system error writing out the state file, but
we'll leave that until later.