Make the options stored as fbuild.options so it won't be cached in env.config
It turns out that it's very convenient to call the build functions using
env.config, as then there's only one instance of the results passed around.
One problem though is that we were previously passing around the options,
and so if you specified "-v" at the command line you'd result in a completely
new build. The easy way to avoid this is to not pass that value using the
ususal function arguments. This has the unfortunate side effect of adding
even more global state, but I haven't figured out a great way to work around
it.