run{,_optional}_tests doesn't make sense any more with the new environment.
Since we now assume that all configuration returns a value, it doesn't make
sense to have run_tests and run_optional_tests, since they didn't return a
value, and there's not a sensible subsitution. So, I'm moving them to the
"simple" test as it's used there to just simply trigger the tests, but not
use any results from it.