Make env a global variable.
Since we use the environment everywhere, and since we've already got a couple
global variables, it'll make life a lot simpler if we just make the env
global. This brings with it a couple other changes. Since we now store more
than just config data in the environment, the function has now been renamed
to cache.

This also comes with a bit of cleanup. It turns out that the c packages
take a builder that acts just like the config value, so we replace that
with just using self.config.

This also fixes some dependency problems with packages.{Copy,Move}.