Add Path.is_dirty, which can determine if a file needs to be rebuilt.
It's turned out that the fbuild.packages hasn't ultimately worked out too
well. The idea was to create a partial-declarative mechanism to describe
the layout of the source, but it ended up just being used in a procedural
manner. So, we'll back up and move the ideas from it into the builders.
The primary tool that we need to bring over is timestamp checking, to make
sure we only rebuild the sources when they've changed. It turns out that
it doesn't require that many changes to switch over. This also has the nice
benefit that we nearly get optimal building for free. The c builders just
need a dependency generator builder.