Add a mechanism to clear the cache if a cached function is changed.
This patch as a very handy feature. Previously, if you modified a cached
function, you had to reconfigure the entire application. This patch adds
the ability to detect if any of the cached functions were changed. If it
does, the cached values will be thrown out and the function will be
reevaluated. It also is able to track dependencies, so if one cached
function A uses values from another cached function B, and B is changed,
we'll recompute both B and A.

This system isn't perfect though. It can't detect if a cached function uses
a non-cached function that changed. Also, there is a case where if you just
modify a constant value we also might not detect that the function changed.
In either case, reconfiguring will use the new functions.