Create a new db backend for fbuild that simplifies caching data.
This is a rewrite of the environment class. It composes of four tables that
track data over multiple runs. First, we track the functions and a hash of
the function's source to handle if the function changed. Second, we track
the arguments to a function call. Third, we track the files and their hash.
Forth, we track the files used in a function call. This database will then
detect when to clear the cache and rerun the function.