build_objects shouldn't be cached.
When a object has extra external dependencies like ocaml files depend on
library changes, we need to add the dependencies in build_objects.
Unfortunately, the db doesn't know about these new files and so it can't
tell when a function really needs to be rerun. So, we'll just not cache
these functions.