Modify the ocaml native build_objects to search for src from libs properly.
As a convience, the directory of a library automatically gets added to the
include path when building. This results in ocamldep not being able to find
some files. The solution is just to add the non-buildroot path to the
includes.

This also tweaks ocamldep to cache the dependencies while we're building a
lib or an exe.