Do not try to stop scheduler when thrown KeyboardInterrupt.
Python threading can be unpredicatable when signaled. Technically, signals
are emitted between atomic python statements. This means that we could be
interrupted in between a finally and a lock.release. This means that we
can't really protect against locks not being properly released. So, if we
are interrupted, just ignore the fact that threads are still going and quit.