Don't use queue.get timeout since it's cpu heavy
Queue's timeout function is pretty heavyweight as
it's trying to minimize the blocking time. This
means that the cpu will spin while it's trying to
get a file. Instead, lets be a little more
wasteful and just do a non-blocking get and if we
fail just sleep for 0.1 seconds. I doubt anyone
will notice the difference.