Pass ids to save_function and save_file.
There was a problem with the sqlite backend where
the ids where being incremented underneath ourselves.
This turned out to be from "INSERT OR REPLACE INTO"
which does a delete then reinsert, which increments
the id. This patch fixes the problem by directly
handling whether or not the record is in the database.